afeguir allotjaments

+++ src/prog2/model/AccesAsfalt.java
@@ -2,14 +2,14 @@ package prog2.model;
 
 public abstract class AccesAsfalt extends Acces{
 
-    private int metresQuadrat;
+    private float metresQuadrat;
 
-    public AccesAsfalt (String nom, boolean obert, int metresQuadrat) {
+    public AccesAsfalt (String nom, boolean obert, float metresQuadrat) {
         super(nom, obert);
         this.metresQuadrat = metresQuadrat;
     }
 
-    public int getMetresQuadrat() {
+    public float getMetresQuadrat() {
         return metresQuadrat;
     }
     public void setMetresQuadrat(int metresQuadrat) {
