+++ src/prog2/model/CamiAsfalt.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiAsfalt extends AccesAsfalt {
 
-    public CamiAsfalt(String nom, boolean obert, int metresQuadrats) {
+    public CamiAsfalt(String nom, boolean obert, float metresQuadrats) {
         super(nom, obert, metresQuadrats);
     }
