+++ src/prog2/model/CamiTerra.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiTerra extends AccesTerra{
 
-    public CamiTerra(String nom, boolean obert, int metres) {
+    public CamiTerra(String nom, boolean obert, float metres) {
         super(nom, obert, metres);
     }
